# server-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: HiGoOS Backend API
  version: 0.1.0
  summary: Contract for the HiGoOS web-pc desktop and first full backend release.
  description: >
    This OpenAPI contract is the source of truth between the Go backend and the
    web-pc frontend. Responses use a common envelope with stable domain schemas.
servers:
  - url: /api/v1
    description: Same-origin API base path.
security:
  - sessionCookie: []
...
paths:
...
  /desktop/apps:
    get:
      tags: [Desktop]
      operationId: listDesktopApps
      summary: List dock applications.
      responses:
        '200':
          $ref: '#/components/responses/DesktopAppsResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /system/backups:
    post:
      tags: [Settings]
      operationId: createSystemBackup
      summary: Create a backup of database, settings, keys, and app state.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SystemBackupRequest'
      responses:
        '200':
          $ref: '#/components/responses/TaskResponse'
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: higo_session
...
  responses:
...
    ErrorResponse:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    DesktopAppsResponse:
      description: Dock apps response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ResponseEnvelope'
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/DesktopApp'
...
    TaskResponse:
      description: Task response.
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ResponseEnvelope'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
...
  schemas:
    ResponseEnvelope:
      type: object
      required: [ok, requestId, data]
      properties:
        ok:
          type: boolean
          const: true
        requestId:
          type: string
        data: {}
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
    ErrorEnvelope:
      type: object
      required: [ok, requestId, error]
      properties:
        ok:
          type: boolean
          const: false
        requestId:
          type: string
        error:
          $ref: '#/components/schemas/Error'
    ResponseMeta:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        source:
          type: string
        eventCursor:
          type: string
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          examples: [permission_denied]
        message:
          type: string
        details:
          type: object
          additionalProperties: true
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    Task:
      type: object
      required: [id, kind, status, progress, createdAt]
      properties:
        id:
          type: string
        kind:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        title:
          type: string
        progress:
          type: number
          minimum: 0
          maximum: 100
        message:
          type: string
        risk:
          $ref: '#/components/schemas/RiskLevel'
        auditId:
          type: string
        result:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    DesktopApp:
      type: object
      required: [id, name, icon]
      properties:
        id:
          type: string
        name:
          type: string
        icon:
          type: string
        badge:
          type: integer
        utility:
          type: boolean
...
    SystemBackupRequest:
      type: object
      properties:
        includeKeys:
          type: boolean
          default: true
        includeAppState:
          type: boolean
          default: true
        destination:
          type: string
...
    TaskStatus:
      type: string
      enum: [queued, running, waiting_confirmation, succeeded, failed, canceled]
...
    RiskLevel:
      type: string
      enum: [low, medium, high]
      description: Frontend may localize as 低风险, 中风险, 高风险.
...